+++ src/ui-actions/conversation-actions.tsx
@@ -3,17 +3,21 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-import { Account, getBridgedFunctions, replaceHistory, t } from '@zextras/carbonio-shell-ui';
-import { forEach, isArray, map } from 'lodash';
 import React from 'react';
+
+import { getBridgedFunctions, replaceHistory, t } from '@zextras/carbonio-shell-ui';
+import { forEach, isArray, map } from 'lodash';
+
+import DeleteConvConfirm from './delete-conv-modal';
+import { errorPage } from './error-page';
+import MoveConvMessage from './move-conv-msg';
 import { getContentForPrint } from '../commons/print-conversation/print-conversation';
 import { ConversationActionsDescriptors } from '../constants';
 import { convAction, getMsgsForPrint } from '../store/actions';
 import { AppDispatch, StoreProvider } from '../store/redux';
+import { ExtraWindowCreationParams, ExtraWindowsContextType } from '../types';
 import type { ConvActionReturnType, Conversation, MailMessage } from '../types';
-import DeleteConvConfirm from './delete-conv-modal';
-import MoveConvMessage from './move-conv-msg';
-import { errorPage } from './error-page';
+import { ConversationPreviewPanel } from '../views/app/detail-panel/conversation-preview-panel';
 
 type ConvActionIdsType = Array<string>;
 type ConvActionValueType = string | boolean;
@@ -56,6 +60,43 @@ export function setConversationsFlag({
 		}
 	};
 }
+
+export const previewConversationOnSeparatedWindow = (
+	conversationId: string,
+	folderId: string,
+	subject: string,
+	createWindow: ExtraWindowsContextType['createWindow']
+): void => {
+	if (!createWindow) {
+		return;
+	}
+
+	const createWindowParams: ExtraWindowCreationParams = {
+		name: `conversation-${conversationId}`,
+		returnComponent: false,
+		children: <ConversationPreviewPanel conversationId={conversationId} folderId={folderId} />,
+		title: subject,
+		closeOnUnmount: false
+	};
+	createWindow(createWindowParams);
+};
+
+export const previewConversationOnSeparatedWindowAction = (
+	conversationId: string,
+	folderId: string,
+	subject: string,
+	createWindow: ExtraWindowsContextType['createWindow']
+): ConvActionReturnType => {
+	const actDescriptor = ConversationActionsDescriptors.PREVIEW_ON_SEPARATED_WINDOW;
+	return {
+		id: actDescriptor.id,
+		icon: 'BrowserOutline',
+		label: t('action.preview_on_separated_window', 'Open on a new window'),
+		onClick: (): void => {
+			previewConversationOnSeparatedWindow(conversationId, folderId, subject, createWindow);
+		}
+	};
+};
 
 export function setMultipleConversationsFlag({
 	ids,
@@ -136,11 +177,9 @@ export function setConversationsRead({
 }
 
 export function printConversation({
-	conversation,
-	account
+	conversation
 }: {
 	conversation: Conversation | Conversation[];
-	account: Account;
 }): ConvActionReturnType {
 	let messageIds: Array<string> = [];
 
@@ -164,7 +203,6 @@ export function printConversation({
 				.then((res) => {
 					const content = getContentForPrint({
 						messages: res,
-						account,
 						conversations: conversation,
 						isMsg: false
 					});