+++ src/views/app/folder-panel/parts/list-item-actions-wrapper.tsx
@@ -3,11 +3,15 @@
  *
  * SPDX-License-Identifier: AGPL-3.0-only
  */
-import { Container, Dropdown, IconButton, Tooltip } from '@zextras/carbonio-design-system';
-import { useTags, useUserAccount } from '@zextras/carbonio-shell-ui';
 import React, { FC, ReactElement, SyntheticEvent, useCallback } from 'react';
+
+import { Container, Dropdown, IconButton, Tooltip } from '@zextras/carbonio-design-system';
+import { useTags } from '@zextras/carbonio-shell-ui';
 import styled, { DefaultTheme } from 'styled-components';
+
+import { isConversation } from '../../../../helpers/messages';
 import { useAppDispatch } from '../../../../hooks/redux';
+import { useMessageActions } from '../../../../hooks/use-message-actions';
 import type {
 	ConvActionReturnType,
 	Conversation,
@@ -17,6 +21,7 @@ import type {
 	TagActionItemType
 } from '../../../../types';
 import { getMsgConvActions } from '../../../../ui-actions/get-msg-conv-actions';
+import { useExtraWindowsManager } from '../../extra-windows/extra-window-manager';
 
 const HoverBarContainer = styled(Container)<{ background: keyof DefaultTheme['palette'] }>`
 	top: 0;
@@ -86,15 +91,17 @@ export const ListItemActionWrapper: FC<ListItemActionWrapperProps> = ({
 	deselectAll
 }) => {
 	const dispatch = useAppDispatch();
-	const account = useUserAccount();
 	const tags = useTags();
+	const { createWindow } = useExtraWindowsManager();
+	const messageActions = useMessageActions(isConversation(item) ? undefined : item, true);
 
 	const [hoverActions, dropdownActions] = getMsgConvActions({
 		item,
 		dispatch,
 		deselectAll,
-		account,
-		tags
+		tags,
+		createWindow,
+		messageActions
 	});
 
 	const dropdownActionsItems = dropdownActions.map((action) => ({