+++ CHANGELOG.md
@@ -2,6 +2,41 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [1.9.2](https://github.com/zextras/carbonio-mails-ui/compare/v1.9.1...v1.9.2) (2023-10-31)
+
+
+### Bug Fixes
+
+* restore message actions on displayer ([8533b8d](https://github.com/zextras/carbonio-mails-ui/commit/8533b8de0dea8b7fc93a9e1561c8807ed70fb840))
+
+### [1.9.1](https://github.com/zextras/carbonio-mails-ui/compare/v1.9.0...v1.9.1) (2023-10-31)
+
+
+### Bug Fixes
+
+* fix function in folders.ts ([630d49c](https://github.com/zextras/carbonio-mails-ui/commit/630d49c5647f82927a4cd84c5996397b462cab88))
+
+## [1.9.0](https://github.com/zextras/carbonio-mails-ui/compare/v1.8.0...v1.9.0) (2023-10-31)
+
+
+### Features
+
+* open messages and conversations in a separated browser tab ([a39d315](https://github.com/zextras/carbonio-mails-ui/commit/a39d315bfb17f7704b064c83ba60b64deeec729e))
+
+## [1.8.0](https://github.com/zextras/carbonio-mails-ui/compare/v1.7.1...v1.8.0) (2023-10-27)
+
+
+### Features
+
+* allow user to copy-click the participants email addresses  ([86ccc03](https://github.com/zextras/carbonio-mails-ui/commit/86ccc03ed68e5aa4a6e1f213a5c0b8460a6dbc7c))
+
+
+### Bug Fixes
+
+* email filter option 'do not process additional filters' is always active ([12540be](https://github.com/zextras/carbonio-mails-ui/commit/12540be7df37e92f1b805b888fedbb570627c938))
+* load conversation full content ([a2653d0](https://github.com/zextras/carbonio-mails-ui/commit/a2653d0229f2fbda2bdcf8650b824d5997a363d9))
+* unnecessary first saving of the draft trigger by the editor ([6ee0f77](https://github.com/zextras/carbonio-mails-ui/commit/6ee0f7766f788a468baf9042d3d6c1ac15513d43))
+
 ### [1.7.1](https://github.com/zextras/carbonio-mails-ui/compare/v1.7.0...v1.7.1) (2023-10-12)
 
 ## [1.7.0](https://github.com/zextras/carbonio-mails-ui/compare/v1.6.16...v1.7.0) (2023-10-03)